pyproject.toml
[project]
name = "claudebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
packages = [{include = "app"}]
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "httpx>=0.28.1",
    "inflection>=0.5.1",
    "psycopg>=3.3.3",
    "pydantic-settings>=2.13.1",
    "python-telegram-bot>=22.6",
    "sqlalchemy[asyncio]>=2.0.47",
]
